TST: Move make_new_server method into BaseTestCase
