+++ tests/resources/test_service.py
@@ -1,6 +1,6 @@
 import concurrent.futures
 import time
-from tests.base import BaseTestCase
+from tests.base import BaseTestCase, MockAnalysis
 
 from octue import exceptions
 from octue.resources.manifest import Manifest
@@ -9,11 +9,6 @@ from octue.resources.service_backend import GCPPubSubBackend
 
 
 SERVER_WAIT_TIME = 5
-
-
-class MockAnalysis:
-    output_values = "Hello! It worked!"
-    output_manifest = None
 
 
 class DifferentMockAnalysis:
@@ -36,14 +31,8 @@ class TestService(BaseTestCase):
     )
     asking_service = Service(name="asker", backend=BACKEND, id="249fc09d-9d6f-45d6-b1a4-0aacba5fca79")
 
-    def make_new_server(self, id="352f8185-1d58-4ddf-8faa-2af96147f96f", run_function=None):
-        """ Make and return a new service ready to serve analyses from its run function. """
-        if not run_function:
-            run_function = lambda input_values, input_manifest: MockAnalysis()  # noqa
-
-        return Service(name=f"server-{id}", backend=self.BACKEND, id=id, run_function=run_function)
-
-    def ask_question_and_wait_for_answer(self, asking_service, responding_service, input_values, input_manifest):
+    @staticmethod
+    def ask_question_and_wait_for_answer(asking_service, responding_service, input_values, input_manifest):
         """ Get an asking service to ask a question to a responding service and wait for the answer. """
         subscription = asking_service.ask(responding_service.id, input_values, input_manifest)
         return asking_service.wait_for_answer(subscription)
@@ -52,7 +41,7 @@ class TestService(BaseTestCase):
         """ Test that a serving service only serves for as long as its timeout. """
         with concurrent.futures.ThreadPoolExecutor(max_workers=3) as executor:
             start_time = time.perf_counter()
-            responding_future = executor.submit(self.make_new_server().serve, timeout=10)
+            responding_future = executor.submit(self.make_new_server(self.BACKEND).serve, timeout=10)
             list(concurrent.futures.as_completed([responding_future]))[0].result()
             self.assertTrue(time.perf_counter() - start_time < 20)
 
@@ -66,7 +55,7 @@ class TestService(BaseTestCase):
 
     def test_ask(self):
         """ Test that a service can ask a question to another service that is serving and receive an answer. """
-        responding_service = self.make_new_server()
+        responding_service = self.make_new_server(self.BACKEND)
 
         with concurrent.futures.ThreadPoolExecutor(max_workers=3) as executor:
             executor.submit(responding_service.serve, timeout=10)
@@ -90,7 +79,7 @@ class TestService(BaseTestCase):
         """ Test that a service can ask a question including an input_manifest to another service that is serving and
         receive an answer.
         """
-        responding_service = self.make_new_server()
+        responding_service = self.make_new_server(self.BACKEND)
 
         with concurrent.futures.ThreadPoolExecutor(max_workers=3) as executor:
             executor.submit(responding_service.serve, timeout=10)
@@ -113,7 +102,7 @@ class TestService(BaseTestCase):
     def test_ask_with_output_manifest(self):
         """ Test that a service can receive an output manifest as part of the answer to a question. """
         responding_service = self.make_new_server(
-            run_function=lambda input_values, input_manifest: MockAnalysisWithOutputManifest()
+            self.BACKEND, run_function=lambda input_values, input_manifest: MockAnalysisWithOutputManifest()
         )
 
         with concurrent.futures.ThreadPoolExecutor(max_workers=3) as executor:
@@ -135,7 +124,7 @@ class TestService(BaseTestCase):
 
     def test_service_can_ask_multiple_questions(self):
         """ Test that a service can ask multiple questions to the same server and expect replies to them all. """
-        responding_service = self.make_new_server()
+        responding_service = self.make_new_server(self.BACKEND)
 
         with concurrent.futures.ThreadPoolExecutor(max_workers=7) as executor:
             executor.submit(responding_service.serve, timeout=10)
@@ -165,10 +154,13 @@ class TestService(BaseTestCase):
     def test_service_can_ask_questions_to_multiple_servers(self):
         """ Test that a service can ask questions to different servers and expect replies to them all. """
         responding_service_1 = self.make_new_server(
-            id="352f8185-1d58-4ddf-8faa-2af96147f96f", run_function=lambda input_values, input_manifest: MockAnalysis()
+            self.BACKEND,
+            id="352f8185-1d58-4ddf-8faa-2af96147f96f",
+            run_function=lambda input_values, input_manifest: MockAnalysis(),
         )
 
         responding_service_2 = self.make_new_server(
+            self.BACKEND,
             id="6a05b695-4aa1-468c-bdf5-73303665165d",
             run_function=lambda input_values, input_manifest: DifferentMockAnalysis(),
         )
